refactor: move auto compare & register function to train.py
